moved widget service into dashboard
